\name{simaCGH}

\alias{simaCGH}

\docType{data}

\title{Simulated aCGH Data}

\description{
  These \eqn{20} samples of aCGH data are simulated using the Fused Lasso Latent
  Feature model with \eqn{5} features, as described in Section 4 of Nowak and
  others (2011).
}

\usage{data(simaCGH)}

\format{
  A matrix consisting of \eqn{1000} probes (rows) and \eqn{20} samples (columns).
}

\references{G. Nowak, T. Hastie, J. R. Pollack and R. Tibshirani.  A
  Fused Lasso Latent Feature Model for Analyzing Multi-Sample aCGH
  Data. \emph{Biostatistics}, 2011, doi: 10.1093/biostatistics/kxr012}

\examples{
data(simaCGH)
}

\keyword{datasets}
